Admin platform identity proof.

+++ hospital_platform/src/main/java/com/hospitalplatform/hospital_platform/controller/HospitalMemberController.java
@@ -3,6 +3,7 @@ package com.hospitalplatform.hospital_platform.controller;
 import com.hospitalplatform.hospital_platform.dto.NewMemberDTO;
 import com.hospitalplatform.hospital_platform.dto.RoleUpdateDTO;
 import com.hospitalplatform.hospital_platform.service.HospitalUserService;
+import com.hospitalplatform.hospital_platform.service.SecurityService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -15,26 +16,44 @@ public class HospitalMemberController {
     @Autowired
     private HospitalUserService hospitalUserService;
 
+    @Autowired
+    private SecurityService securityService;
+
     @PostMapping
-    public ResponseEntity<?> createMember(@RequestBody @Validated NewMemberDTO member) throws Exception {
+    public ResponseEntity<?> createMember(@RequestHeader("X-sudo-key") String key,
+                                          @RequestBody @Validated NewMemberDTO member) throws Exception {
+        if (!securityService.checkSecurityToken(key))
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+
         hospitalUserService.createUser(member);
         return new ResponseEntity(null, HttpStatus.ACCEPTED);
     }
 
     @DeleteMapping("/{memberId}")
-    public ResponseEntity<?>  deleteMember(@PathVariable Integer memberId) {
+    public ResponseEntity<?>  deleteMember(@RequestHeader("X-sudo-key") String key,
+                                           @PathVariable Integer memberId) {
+        if (!securityService.checkSecurityToken(key))
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+
         hospitalUserService.deleteUser(memberId);
         return new ResponseEntity<>(null, HttpStatus.NO_CONTENT);
     }
 
     @GetMapping
-    public ResponseEntity<?> getMembers() {
+    public ResponseEntity<?> getMembers(@RequestHeader("X-sudo-key") String key) {
+        if (!securityService.checkSecurityToken(key))
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+
         return new ResponseEntity<>(hospitalUserService.getHospitalUsers(), HttpStatus.OK);
     }
 
     @PutMapping("/{memberId}/roles")
-    public ResponseEntity<?> changeMemberRole(
-            @PathVariable Integer memberId, @RequestBody @Validated RoleUpdateDTO roleUpdateDTO) {
+    public ResponseEntity<?> changeMemberRole(@RequestHeader("X-sudo-key") String key,
+                                              @PathVariable Integer memberId,
+                                              @RequestBody @Validated RoleUpdateDTO roleUpdateDTO) {
+        if (!securityService.checkSecurityToken(key))
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+
         hospitalUserService.changeUserRole(memberId, roleUpdateDTO);
         return new ResponseEntity<>(null, HttpStatus.ACCEPTED);
     }
